Access PhysLoc by src/dst index rather than SSATmp.
Right now, CodeGenerator functions get PhysLoc indirectly
by: reginfo[inst.id][inst->src(i).id].  There's no need for
the SSATmp being the key in the second dimension, and its
more expensive anyway since the second map is sparse.
This diff refactors RegAllocInfo so the second dimension
is an array indexed by src or dst index:
reginfo[inst.id].src[i]
reginfo[inst.id].dst[i]

The actual PhysLocs are stored internally in one vector
per instruction (dsts, then srcs).  Now, both dimensions
are vectors, so lookups should be faster. It was necessary
to plumb PhysLocs through several helper functions, but
also it eliminated accessing SSATmps in several cases.

This diff also prepares for per-arch temp operands:
reginfo[inst.id].tmp[i]

Reviewed By: @swtaarrs

Differential Revision: D1173989

+++ hphp/runtime/vm/jit/phys-reg.h
@@ -223,7 +223,7 @@ private:
   friend struct RegSet;
   explicit constexpr PhysReg(int n) : n(n) {}
 
-  int n;
+  int8_t n;
 };
 
 constexpr PhysReg InvalidReg;
